Refactor demo loop to have a Runner
This will be further cleaned up after we remove the diagnostic classes.
