+++ app/demo-loop/Transporter.hh
@@ -8,25 +8,24 @@
 #pragma once
 
 #include <memory>
-#include <string>
-#include <unordered_map>
 #include <utility>
 #include <vector>
 
 #include "corecel/Assert.hh"
 #include "corecel/Types.hh"
-#include "corecel/cont/Range.hh"
 #include "corecel/cont/Span.hh"
-#include "corecel/math/NumericLimits.hh"
+#include "corecel/sys/ThreadId.hh"
 #include "celeritas/Types.hh"
-#include "celeritas/global/CoreParams.hh"
-#include "celeritas/phys/Primary.hh"
+
+#include "Runner.hh"
+#include "RunnerInput.hh"
 
 namespace celeritas
 {
 struct Primary;
 template<MemSpace M>
 class Stepper;
+class CoreParams;
 }
 
 namespace demo_loop
@@ -34,18 +33,7 @@ namespace demo_loop
 //---------------------------------------------------------------------------//
 template<celeritas::MemSpace M>
 class Diagnostic;
-
-//---------------------------------------------------------------------------//
-struct EnergyDiagInput
-{
-    using size_type = celeritas::size_type;
-    using real_type = celeritas::real_type;
-
-    char axis{'z'};
-    real_type min{-700};
-    real_type max{700};
-    size_type num_bins{1024};
-};
+struct RunnerInput;
 
 //---------------------------------------------------------------------------//
 //! Input parameters to the transporter.
@@ -53,12 +41,6 @@ struct TransporterInput
 {
     using size_type = celeritas::size_type;
     using CoreParams = celeritas::CoreParams;
-
-    //! Arbitrarily high number for not stopping the simulation short
-    static constexpr size_type no_max_steps()
-    {
-        return celeritas::numeric_limits<size_type>::max();
-    }
 
     // Stepper input
     std::shared_ptr<CoreParams const> params;
@@ -72,7 +54,6 @@ struct TransporterInput
     bool enable_diagnostics{true};
     EnergyDiagInput energy_diag;
 
-    // Threading (TODO)
     celeritas::StreamId stream_id{0};
 
     //! True if all params are assigned
@@ -80,45 +61,6 @@ struct TransporterInput
     {
         return params && num_track_slots > 0 && max_steps > 0;
     }
-};
-
-//---------------------------------------------------------------------------//
-//! Simulation timing results.
-struct TransporterTiming
-{
-    using real_type = celeritas::real_type;
-    using VecReal = std::vector<real_type>;
-    using MapStrReal = std::unordered_map<std::string, real_type>;
-
-    VecReal steps;  //!< Real time per step
-    real_type total{};  //!< Total simulation time
-    real_type setup{};  //!< One-time initialization cost
-    MapStrReal actions{};  //!< Accumulated action timing
-};
-
-//---------------------------------------------------------------------------//
-//! Tallied result and timing from transporting a set of primaries
-struct TransporterResult
-{
-    //!@{
-    //! \name Type aliases
-    using real_type = celeritas::real_type;
-    using size_type = celeritas::size_type;
-    using VecCount = std::vector<size_type>;
-    using VecReal = std::vector<real_type>;
-    using MapStringCount = std::unordered_map<std::string, size_type>;
-    using MapStringVecCount = std::unordered_map<std::string, VecCount>;
-    //!@}
-
-    //// DATA ////
-
-    VecCount initializers;  //!< Num starting track initializers
-    VecCount active;  //!< Num tracks active at beginning of step
-    VecCount alive;  //!< Num living tracks at end of step
-    VecReal edep;  //!< Energy deposition along the grid
-    MapStringCount process;  //!< Count of particle/process interactions
-    MapStringVecCount steps;  //!< Distribution of steps
-    TransporterTiming time;  //!< Timing information
 };
 
 //---------------------------------------------------------------------------//
@@ -151,6 +93,7 @@ class TransporterBase
     using SpanConstPrimary = celeritas::Span<const celeritas::Primary>;
     using CoreParams = celeritas::CoreParams;
     using ActionId = celeritas::ActionId;
+    using TransporterResult = RunnerResult;
     //!@}
 
   public:
